+++ src/main/java/frc/robot/commands/Autos.java
@@ -17,7 +17,7 @@ import frc.robot.Constants.LiftConstants;
 import frc.robot.commands.drive.DriveDistance;
 import frc.robot.commands.drive.RotateToAngle;
 import frc.robot.commands.gripper.Hold;
-import frc.robot.subsystems.AddressableLEDSubsystem;
+import frc.robot.subsystems.LEDSystem;
 import frc.robot.subsystems.DriveSystem;
 import frc.robot.subsystems.GripperSystem;
 import frc.robot.subsystems.LiftSystem;
